refactor: extract sandbox into separate package
