+++ apps/cli/package.json
@@ -15,11 +15,10 @@
   },
   "dependencies": {
     "@open-harness/agent": "workspace:*",
+    "@open-harness/sandbox": "workspace:*",
     "@open-harness/tui": "workspace:*",
-    "@vercel/sandbox": "^1.1.2",
     "ink": "^6.6.0",
     "ink-spinner": "^5.0.0",
-    "just-bash": "^2.1.0",
     "react": "^19.2.3"
   },
   "devDependencies": {
