+++ packages/agent/subagents/executor.ts
@@ -5,7 +5,7 @@ import { writeFileTool, editFileTool } from "../tools/write";
 import { grepTool } from "../tools/grep";
 import { globTool } from "../tools/glob";
 import { bashTool, commandNeedsApproval } from "../tools/bash";
-import type { Sandbox } from "../sandbox";
+import type { Sandbox } from "@open-harness/sandbox";
 
 const EXECUTOR_SYSTEM_PROMPT = `You are an executor agent - a fire-and-forget subagent that completes specific, well-defined implementation tasks autonomously.
