+++ packages/sandbox/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@open-harness/tsconfig/bun.json",
+  "include": ["./**/*"],
+  "exclude": ["node_modules"]
+}
